requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
idna==2.7
Markdown==2.6.11
oauthlib==2.1.0
PyJWT==1.6.4
python-dotenv==0.8.2
python3-openid==3.1.0
pytz==2018.4
requests==2.19.0
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
